+++ docs/source/conf.py
@@ -41,7 +41,7 @@ extensions = [
     'sphinx.ext.mathjax',
     'sphinxarg.ext',
     #'sphinxcontrib.bibtex',
-    'sphinxcontrib.programoutput',
+    #'sphinxcontrib.programoutput', # unused, argparse used instead
 ]
 
 # additional themes
@@ -96,7 +96,7 @@ copyright = u'2013-2015, Francesco Rubino'
 #
 # The short X.Y version.
 
-__VERSION__ = '0.3.2'
+__VERSION__ = '0.3.3'
 
 version = __VERSION__
 # The full version, including alpha/beta/rc tags.